Introduce FullTextSearchRule as abstraction of ContainBasedSearchRule and RegexBasedSearchRule
