+++ src/main/java/org/jabref/model/search/rules/RegexBasedSearchRule.java
@@ -1,23 +1,15 @@
 package org.jabref.model.search.rules;
 
-import java.io.IOException;
 import java.util.EnumSet;
-import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
-import java.util.stream.Collectors;
 
 import org.jabref.architecture.AllowedToUseLogic;
-import org.jabref.gui.Globals;
-import org.jabref.logic.pdf.search.retrieval.PdfSearcher;
-import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.Field;
-import org.jabref.model.pdf.search.PdfSearchResults;
-import org.jabref.model.pdf.search.SearchResult;
 import org.jabref.model.search.rules.SearchRules.SearchFlags;
 
 import org.slf4j.Logger;
@@ -27,25 +19,12 @@ import org.slf4j.LoggerFactory;
  * Search rule for regex-based search.
  */
 @AllowedToUseLogic("Because access to the lucene index is needed")
-public class RegexBasedSearchRule implements SearchRule {
+public class RegexBasedSearchRule extends FullTextSearchRule {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(GrammarBasedSearchRule.class);
 
-    private final EnumSet<SearchFlags> searchFlags;
-
-    private String lastQuery;
-    private List<SearchResult> lastSearchResults;
-
-    private final BibDatabaseContext databaseContext;
-
     public RegexBasedSearchRule(EnumSet<SearchFlags> searchFlags) {
-        this.searchFlags = searchFlags;
-
-        databaseContext = Globals.stateManager.getActiveDatabase().orElse(null);
-    }
-
-    public EnumSet<SearchFlags> getSearchFlags() {
-        return searchFlags;
+        super(searchFlags);
     }
 
     @Override
@@ -86,24 +65,4 @@ public class RegexBasedSearchRule implements SearchRule {
         return getFulltextResults(query, bibEntry).numSearchResults() > 0;
     }
 
-    @Override
-    public PdfSearchResults getFulltextResults(String query, BibEntry bibEntry) {
-
-        if (!searchFlags.contains(SearchRules.SearchFlags.FULLTEXT) || databaseContext == null) {
-            return new PdfSearchResults(List.of());
-        }
-
-        if (!query.equals(this.lastQuery)) {
-            this.lastQuery = query;
-            lastSearchResults = List.of();
-            try {
-                PdfSearcher searcher = PdfSearcher.of(databaseContext);
-                PdfSearchResults results = searcher.search(query, 5);
-                lastSearchResults = results.getSortedByScore();
-            } catch (IOException e) {
-                LOGGER.error("Could not retrieve search results!", e);
-            }
-        }
-        return new PdfSearchResults(lastSearchResults.stream().filter(searchResult -> searchResult.isResultFor(bibEntry)).collect(Collectors.toList()));
-    }
 }
